server/requirements.txt
Django==2.1.*
requests
pillow
pykew